updated placeBet and separated different APIs in different files
