+++ src/controller/gameController/index.js
@@ -2,10 +2,7 @@ import { getGameById } from "./getGamebyId.js";
 import { getGames } from "./getGames.js";
 import { getCurrentGame } from "./getCurrentGame.js";
 import { getGameHistory } from "./getGameHistory.js";
-import {
-  placeBet,
-  getValidBetOptions,
-} from "../gameController/Betting/placeBet.js";
+import { placeBet } from "../gameController/Betting/placeBet.js";
 import { getBettingRange } from "./Betting/getBettingRange.js";
 import { getGamesByCategory } from "./getGamesByCategory.js";
 import { getWinningCards } from "./getWinningCards.js";
@@ -16,7 +13,6 @@ export {
   getGameById,
   getGames,
   placeBet,
-  getValidBetOptions,
   getBettingRange,
   getGameHistory,
   getCurrentGame,
